% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brleur.R
\name{brl_eur}
\alias{brl_eur}
\title{Provides a plot of the Brazilian Real/Euro Exchange Rate}
\usage{
brl_eur()
}
\arguments{
\item{-}{requires no paramenters}
}
\value{
ggplot graph
}
\description{
Provides a plot of the Brazilian Real/Euro Exchange Rate
}
\examples{
brl_eur()

}
